Support directory symlinks in setup_all

+++ setup_all.sh
@@ -18,7 +18,13 @@ for i in `find \`pwd\` -iname '*.symlink'`; do
   homefile=$HOME/$file
   if [ ! -L $homefile ]; then
     echo "• $i -> $homefile"
-    ln -s -i $i $homefile
+    if [[ -d $homefile && -d $i ]]; then
+      echo "Directory $homefile exists so cannot be changed to a symlink."
+      echo "Please remove and re-run setup if you want to use the .dotfiles version."
+      read -rsp $'Press any key to continue...\n' -n1 key
+    else
+      ln -s -i $i $homefile
+    fi
   else
     echo "• $homefile is already a symlink"
   fi
